Fixed GitIgnore, Menu finished, resources
Fixed the git.ignore since it was deleting the proyect configurations as linker settings and others.
finished most of the menu details and added to it a parallax.
added resources such as parallax, ship, nave, credits.txt, etc.
added code to screen mannager from another project (yet to set into this project).